Implement ?ganyu and ?zhongli
Separated meme_commands to a separate folder with separate file
processing to avoid clutter of multiple meme files with the current
one command per file scheme
